Bumped version to v1.2.2 Added ScriptablePluginEngine.getPluginInstance() to resolve a scriptable plugin instance.
