+++ ScriptableMC-Engine-JS/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: ScriptableMC-Engine-JS
 description: 'Run JavaScript/TypeScript plugins for Bukkit/Spigot/Paper Minecraft 1.15'
-version: '1.2.1'
+version: '1.2.2'
 api-version: '1.13'
 author: AStorks
 main: com.pixlfox.scriptablemc.SMCJavaScriptEngineMain
