+++ ScriptableMC-Engine-PY/src/main/kotlin/com/pixlfox/scriptablemc/core/PythonPluginEngine.kt
@@ -2,7 +2,6 @@ package com.pixlfox.scriptablemc.core
 
 import com.pixlfox.scriptablemc.SMCPythonConfig
 import com.pixlfox.scriptablemc.ScriptEngineMain
-import com.smc.exceptions.ScriptNotFoundException
 import com.pixlfox.scriptablemc.utils.UnzipUtility
 import fr.minuskube.inv.InventoryManager
 import org.graalvm.polyglot.*
@@ -53,41 +52,6 @@ class PythonPluginEngine(override val bootstrapPlugin: ScriptEngineMain, overrid
         globalBindings = graalContext.getBindings(languageName)
     }
 
-    override fun loadMainScript(path: String) {
-        try {
-            val mainScriptFile = File(path)
-            if(!mainScriptFile.parentFile.exists()) {
-                mainScriptFile.parentFile.mkdirs()
-            }
-
-            if(mainScriptFile.exists()) {
-                val mainReturn = eval(
-                    Source.newBuilder(languageName, mainScriptFile)
-                        .name(mainScriptFile.name)
-                        .interactive(false)
-                        .build()
-                )
-
-                // Load all plugin types returned as an array
-                if(mainReturn.hasArrayElements()) {
-                    for (i in 0 until mainReturn.arraySize) {
-                        this.loadPlugin(mainReturn.getArrayElement(i))
-                    }
-
-                    // Enable all plugins if not already enabled
-                    if(!enabledAllPlugins) {
-                        enableAllPlugins()
-                    }
-                }
-            }
-            else {
-                throw ScriptNotFoundException(mainScriptFile)
-            }
-        }
-        catch(ex: Exception) {
-            startupErrors.add(ex)
-        }
-    }
 
     override fun start() {
         instance = this
