+++ ScriptableMC-Engine-PY/src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: ScriptableMC-Engine-PY
 description: 'Run Python plugins for Bukkit/Spigot/Paper Minecraft 1.15'
-version: '1.2.1'
+version: '1.2.2'
 api-version: '1.13'
 author: AStorks
 main: com.pixlfox.scriptablemc.SMCPythonEngineMain
