+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 
 allprojects {
     group = "com.pixlfox.scriptablemc"
-    version = "1.2.1"
+    version = "1.2.2"
 
     repositories {
         mavenCentral()
